requirements.txt
annotated-types==0.6.0
anthropic==0.19.1
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.87
botocore==1.34.87
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
cssselect==1.2.0
distro==1.9.0
fake-useragent==1.5.1
filelock==3.13.1
firebase-admin==6.4.0
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.1.0
itemadapter==0.8.0
jmespath==1.0.1
json5==0.9.14
lxml==5.1.0
msgpack==1.0.7
openai==1.12.0
outcome==1.3.0.post0
packaging==23.2
parse==1.20.1
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyee==11.1.0
PyJWT==2.8.0
pyparsing==3.1.1
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
rsa==4.9
s3transfer==0.10.1
scrapeops_python_logger==0.4.7
scrapeops_python_requests==0.4.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tld==0.13
tokenizers==0.15.2
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.18.2
